+++ include/rocksdb/options.h
@@ -20,13 +20,13 @@
 #include "rocksdb/advanced_options.h"
 #include "rocksdb/comparator.h"
 #include "rocksdb/compression_type.h"
+#include "rocksdb/data_structure.h"
 #include "rocksdb/env.h"
 #include "rocksdb/file_checksum.h"
 #include "rocksdb/listener.h"
 #include "rocksdb/sst_partitioner.h"
 #include "rocksdb/types.h"
 #include "rocksdb/universal_compaction.h"
-#include "rocksdb/utilities/data_structure.h"
 #include "rocksdb/version.h"
 #include "rocksdb/write_buffer_manager.h"
