+++ table/sst_file_writer.cc
@@ -182,10 +182,11 @@ SstFileWriter::~SstFileWriter() {
 Status SstFileWriter::Open(const std::string& file_path) {
   Rep* r = rep_.get();
   Status s;
-  std::unique_ptr<WritableFile> sst_file;
+  std::unique_ptr<FSWritableFile> sst_file;
   FileOptions cur_file_opts(r->env_options);
   cur_file_opts.handoff_checksum_type = ChecksumType::kCRC32c;
-  s = r->ioptions.env->NewWritableFile(file_path, &sst_file, cur_file_opts);
+  s = r->ioptions.env->GetFileSystem()->NewWritableFile(
+      file_path, r->env_options, &sst_file, nullptr);
   if (!s.ok()) {
     return s;
   }
@@ -259,8 +260,8 @@ Status SstFileWriter::Open(const std::string& file_path) {
       0 /* file_creation_time */, "SST Writer" /* db_id */, db_session_id);
   FileTypeSet tmp_set = r->ioptions.checksum_handoff_file_types;
   r->file_writer.reset(new WritableFileWriter(
-      NewLegacyWritableFileWrapper(std::move(sst_file)), file_path,
-      cur_file_opts, r->ioptions.env, nullptr /* io_tracer */,
+      std::move(sst_file), file_path, r->env_options,
+      r->ioptions.env->GetSystemClock(), nullptr /* io_tracer */,
       nullptr /* stats */, r->ioptions.listeners,
       r->ioptions.file_checksum_gen_factory,
       tmp_set.Contains(FileType::kTableFile)));
